Remove the dependency on log4j version 1
The reason for this removal is the following:

A security vulnerability, CVE-2019-17571 has been identified against Log4j 1. Log4j includes a SocketServer that accepts serialized log events and deserializes them without verifying whether the objects are allowed or not. This can provide an attack vector that can be exploited. Since Log4j 1 is no longer maintained, this issue will not be fixed.

log4j 1 was only included for the Log facade. As this log facade is not involved during the generation and is not shipped with the generated API, this removal is harmless.