SqlClientHelper: fix transaction management
Related to https://github.com/quarkusio/quarkus/issues/16430

When a problem occurs on the DB side (e.g. constraint violation), the SqlClientHelper implementation must convey the original error message to the user.

Also, in order to avoid a redundant call to the SqlClient#rollback, we can check if the failure is a TransactionRollbackException.