ClassCastException thrown by stream methods on Iterable types
See #1188

The exception is thrown if the iterable type argument is a Vert.x API type.

In this case, the type transformation is not applied (it's applied correctly for iterator methods).

The fix consists in generating a stream method that uses the iterator from the generated type, instead of invoking the delegate directly.